Move xiaomi_miio base entity to separate module
