+++ homeassistant/components/xiaomi_miio/air_quality.py
@@ -18,7 +18,7 @@ from .const import (
     MODEL_AIRQUALITYMONITOR_S1,
     MODEL_AIRQUALITYMONITOR_V1,
 )
-from .device import XiaomiMiioEntity
+from .entity import XiaomiMiioEntity
 
 _LOGGER = logging.getLogger(__name__)
